refactor: DragIcon
